refactored lightbox into separate components; fixed some thumbnails along the bottom of the lightbox view
