--- src/photos/components/Lightbox.tsx
@@ -1,218 +1,0 @@
-import {
-    ArrowDownTrayIcon,
-    ArrowTopRightOnSquareIcon,
-    ArrowUturnLeftIcon,
-    ChevronLeftIcon,
-    ChevronRightIcon,
-    XMarkIcon,
-} from '@heroicons/react/24/outline'
-import { variants } from '@src/photos/utils/animationVariants'
-import downloadPhoto from '@src/photos/utils/downloadPhoto'
-import { range } from '@src/photos/utils/range'
-import type { ImageProps, SharedModalProps } from '@src/photos/utils/types'
-import { AnimatePresence, MotionConfig, motion } from 'framer-motion'
-import Image from 'next/image'
-import { useState } from 'react'
-import { useSwipeable } from 'react-swipeable'
-
-export default function Lightbox({
-    index,
-    images,
-    changePhotoId,
-    closeModal,
-    navigation,
-    currentPhoto,
-    direction,
-    gallery,
-}: SharedModalProps) {
-    const [loaded, setLoaded] = useState(false)
-    const filteredImages = images?.filter((img: ImageProps) => range(index - 15, index + 15).includes(img.id))
-    const handlers = useSwipeable({
-        onSwipedLeft: () => {
-            if (images?.length && images?.length > 0 && index < images.length - 1) {
-                changePhotoId(index + 1)
-            }
-        },
-        onSwipedRight: () => {
-            if (index > 0) {
-                changePhotoId(index - 1)
-            }
-        },
-        trackMouse: true,
-    })
-    const currentImage = images ? images[index] : currentPhoto
-    const { title } = gallery
-    // console.log('<SharedModal> ', { images, index, navigation, loaded })
-
-    return (
-        <MotionConfig
-            transition={{
-                x: { type: 'spring', stiffness: 300, damping: 30 },
-                opacity: { duration: 0.2 },
-            }}
-        >
-            <div
-                className="relative z-50 flex aspect-[3/2] w-full max-w-7xl items-center wide:h-full xl:taller-than-854:h-auto"
-                {...handlers}
-            >
-                {/* Main image */}
-                <div className="w-full overflow-hidden">
-                    <div className="relative flex aspect-[3/2] items-center justify-center">
-                        <AnimatePresence initial={false} custom={direction}>
-                            <motion.div
-                                key={index}
-                                custom={direction}
-                                variants={variants}
-                                initial="enter"
-                                animate="center"
-                                exit="exit"
-                                className="absolute"
-                            >
-                                {currentImage && currentImage.resource_type === 'image' ? (
-                                    <Image
-                                        src={`https://res.cloudinary.com/${
-                                            process.env.NEXT_PUBLIC_CLOUDINARY_CLOUD_NAME
-                                        }/image/upload/c_scale,${navigation ? 'w_1280' : 'w_1920'}/${
-                                            currentImage.public_id
-                                        }.${currentImage.format}`}
-                                        width={navigation ? 1280 : 1920}
-                                        height={navigation ? 853 : 1280}
-                                        priority
-                                        alt={`${title} image`}
-                                        onLoadingComplete={() => setLoaded(true)}
-                                    />
-                                ) : currentImage ? (
-                                    // eslint-disable-next-line jsx-a11y/media-has-caption
-                                    <video
-                                        src={currentImage.secure_url}
-                                        controls={true}
-                                        width={currentImage.width}
-                                        height={currentImage.height}
-                                        onCanPlay={() => {
-                                            // console.log('onCanPlay')
-                                            setLoaded(true)
-                                        }}
-                                    />
-                                ) : (
-                                    <p>No photo!</p>
-                                )}
-                            </motion.div>
-                        </AnimatePresence>
-                    </div>
-                </div>
-
-                {/* Buttons + bottom nav bar */}
-                <div className="absolute inset-0 mx-auto flex max-w-7xl items-center justify-center">
-                    {/* Buttons */}
-                    {loaded ? (
-                        <div className="relative aspect-[3/2] max-h-full w-full">
-                            {navigation ? (
-                                <>
-                                    {index > 0 && (
-                                        <button
-                                            className="absolute left-3 top-[calc(50%-16px)] rounded-full bg-black/50 p-3 text-white/75 backdrop-blur-lg transition hover:bg-black/75 hover:text-white focus:outline-none"
-                                            style={{ transform: 'translate3d(0, 0, 0)' }}
-                                            onClick={() => changePhotoId(index - 1)}
-                                        >
-                                            <ChevronLeftIcon className="h-6 w-6" />
-                                        </button>
-                                    )}
-                                    {images && index + 1 < images.length ? (
-                                        <button
-                                            className="absolute right-3 top-[calc(50%-16px)] rounded-full bg-black/50 p-3 text-white/75 backdrop-blur-lg transition hover:bg-black/75 hover:text-white focus:outline-none"
-                                            style={{ transform: 'translate3d(0, 0, 0)' }}
-                                            onClick={() => changePhotoId(index + 1)}
-                                        >
-                                            <ChevronRightIcon className="h-6 w-6" />
-                                        </button>
-                                    ) : null}
-                                </>
-                            ) : null}
-                            <div className="absolute top-0 right-0 flex items-center gap-2 p-3 text-white">
-                                {navigation ? (
-                                    <a
-                                        href={`https://res.cloudinary.com/${process.env.NEXT_PUBLIC_CLOUDINARY_CLOUD_NAME}/image/upload/${currentImage?.public_id}.${currentImage?.format}`}
-                                        className="rounded-full bg-black/50 p-2 text-white/75 backdrop-blur-lg transition hover:bg-black/75 hover:text-white"
-                                        target="_blank"
-                                        title="Open fullsize version"
-                                        rel="noreferrer"
-                                    >
-                                        <ArrowTopRightOnSquareIcon className="h-5 w-5" />
-                                    </a>
-                                ) : null}
-                                <button
-                                    onClick={() =>
-                                        downloadPhoto(
-                                            `https://res.cloudinary.com/${process.env.NEXT_PUBLIC_CLOUDINARY_CLOUD_NAME}/image/upload/${currentImage?.public_id}.${currentImage?.format}`,
-                                            `${index}.jpg`
-                                        )
-                                    }
-                                    className="rounded-full bg-black/50 p-2 text-white/75 backdrop-blur-lg transition hover:bg-black/75 hover:text-white"
-                                    title="Download fullsize version"
-                                >
-                                    <ArrowDownTrayIcon className="h-5 w-5" />
-                                </button>
-                            </div>
-                            <div className="absolute top-0 left-0 flex items-center gap-2 p-3 text-white">
-                                <button
-                                    onClick={() => closeModal()}
-                                    className="rounded-full bg-black/50 p-2 text-white/75 backdrop-blur-lg transition hover:bg-black/75 hover:text-white"
-                                >
-                                    {navigation ? (
-                                        <XMarkIcon className="h-5 w-5" />
-                                    ) : (
-                                        <ArrowUturnLeftIcon className="h-5 w-5" />
-                                    )}
-                                </button>
-                            </div>
-                        </div>
-                    ) : null}
-                    {/* Bottom Nav bar */}
-                    {navigation ? (
-                        <div className="fixed inset-x-0 bottom-0 z-40 overflow-hidden bg-gradient-to-b from-black/0 to-black/60">
-                            <motion.div initial={false} className="mx-auto mt-6 mb-6 flex aspect-[3/2] h-14">
-                                <AnimatePresence initial={false}>
-                                    {filteredImages
-                                        ? filteredImages.map(({ public_id, format, id }) => (
-                                              <motion.button
-                                                  initial={{
-                                                      width: '0%',
-                                                      x: `${Math.max((index - 1) * -100, 15 * -100)}%`,
-                                                  }}
-                                                  animate={{
-                                                      scale: id === index ? 1.25 : 1,
-                                                      width: '100%',
-                                                      x: `${Math.max(index * -100, 15 * -100)}%`,
-                                                  }}
-                                                  exit={{ width: '0%' }}
-                                                  onClick={() => changePhotoId(id)}
-                                                  key={id}
-                                                  className={`${
-                                                      id === index ? 'z-20 rounded-md shadow shadow-black/50' : 'z-10'
-                                                  } ${id === 0 ? 'rounded-l-md' : ''} ${
-                                                      images && id === images.length - 1 ? 'rounded-r-md' : ''
-                                                  } relative inline-block w-full shrink-0 transform-gpu overflow-hidden focus:outline-none`}
-                                              >
-                                                  <Image
-                                                      alt="small photos on the bottom"
-                                                      width={180}
-                                                      height={120}
-                                                      className={`${
-                                                          id === index
-                                                              ? 'brightness-110 hover:brightness-110'
-                                                              : 'brightness-50 contrast-125 hover:brightness-75'
-                                                      } h-full transform object-cover transition`}
-                                                      src={`https://res.cloudinary.com/${process.env.NEXT_PUBLIC_CLOUDINARY_CLOUD_NAME}/image/upload/c_scale,w_180/${public_id}.${format}`}
-                                                  />
-                                              </motion.button>
-                                          ))
-                                        : null}
-                                </AnimatePresence>
-                            </motion.div>
-                        </div>
-                    ) : null}
-                </div>
-            </div>
-        </MotionConfig>
-    )
-}
